+++ src/interfaces/ITextRendering.ts
@@ -2,7 +2,7 @@ import { Context } from '../utils/types'
 import ISimplePoint from './simple/ISimplePoint'
 
 export default interface ITextRendering {
-    updateText(ctx: Context, point: ISimplePoint, text: string[]): string[]
+    updateText(ctx: Context, point: ISimplePoint, preUpdatedText: string[]): string[]
     strokeDrawText(ctx: Context, args: [string, number, number]): void
     fillDrawText(ctx: Context, args: [string, number, number]): void
 }