+++ src/objects/EventEmitter.ts
@@ -1,36 +1,35 @@
 import { Dict, EventCallback } from "../utils/types"
 import IEventOptions from "../interfaces/options/IEventOptions"
 import CustomObject from "./core/CustomObject"
-import ValueEvent from '../utils/events/ValueEvent'
 
-export default class EventEmitter<T extends Event = ValueEvent, K extends string = string> extends CustomObject {
-    protected _events: Dict<Set<EventCallback<T>>>
+export default class EventEmitter<Type = any, Key extends string = string> extends CustomObject {
+    protected _events: Dict<Set<EventCallback<Type>>>
 
     constructor() {
         super()
         this._events = new Map()
     }
 
-    on(key: K, callback: EventCallback<T>, {isOnce}: IEventOptions = {}): void {
+    on(key: Key, callback: EventCallback<Type>, {isOnce}: IEventOptions = {}): void {
         if(!this._events.has(key)) this._events.set(key, new Set())
 
-        let eventCallback: EventCallback<T> = callback
+        let eventCallback: EventCallback<Type> = callback
         if(isOnce) eventCallback = e => {
             callback(e)
             this.off(key, eventCallback)
         }
 
-        let set: Set<EventCallback<T>> = this._events.get(key)!
+        let set: Set<EventCallback<Type>> = this._events.get(key)!
         
         if (set.has(eventCallback)) return
         set.add(eventCallback)
     }
 
-    once(key: K, callback: EventCallback<T>): void {
+    once(key: Key, callback: EventCallback<Type>): void {
         this.on(key, callback, {isOnce: true})
     }
 
-    off(key: K, callback: EventCallback<T>): boolean {
+    off(key: Key, callback: EventCallback<Type>): boolean {
         let set = this._events.get(key)
 
         if(!set) return false
@@ -38,19 +37,12 @@ export default class EventEmitter<T extends Event = ValueEvent, K extends string
         return set.delete(callback)
     }
 
-    emit(event: T): void {
-        let { type } = event
-        
-        let set = this._events.get(type)
+    emit(key: Key, value?: Type): void {
+        let set = this._events.get(key)
         if(!set) return
 
         set.forEach(cb => {
-            cb(event)
+            cb(value)
         })
-    }
-
-    emitDefault(key: K): void {
-        const event = new Event(key) as T
-        this.emit(event)
     }
 }