+++ src/utils/events/ValueEvent.ts
@@ -1,3 +1,0 @@
-export default class ValueEvent<T = any> extends Event {
-    public value?: T
-}